+++ CHANGELOG.md
@@ -4,5 +4,6 @@
 
 #### 25.10.2018
 
+* [General] Video player 
 * [General] Project & repository setup